8.10

+++ Templates/FastPower.cpp
@@ -2,9 +2,9 @@
 快速幂
 PS: 有a*a运算，有溢出的可能
 */
-int power(long long a, int n, int m)
+ll power(ll a, ll n, ll m)
 {
-	long long ans = 1;
+	ll ans = 1;
 	while (n)
 	{
 		if (n & 1) ans = ans * a % m;
